+++ demo/src/features/text/internal/mod.rs
@@ -1,0 +1,3 @@
+mod font_atlas_cache;
+pub use font_atlas_cache::FontAtlasCache;
+pub use font_atlas_cache::TextImageUpdate;
