+++ demo/src/features/text/prepare.rs
@@ -1,41 +1,43 @@
-use super::write::TextCommandWriter;
-use crate::features::text::write::TextDrawCallBuffers;
-use crate::features::text::{
-    ExtractedTextData, FontAtlasCache, TextRenderFeature, TextUniformBufferObject,
-};
+rafx::declare_render_feature_prepare_job!();
+
+use super::internal::FontAtlasCache;
+use super::public::TextDrawCommand;
+use super::TextUniformBufferObject;
+use crate::assets::font::FontAsset;
 use crate::phases::UiRenderPhase;
+use distill::loader::LoadHandle;
+use fnv::FnvHashMap;
 use rafx::api::RafxBufferDef;
 use rafx::framework::{MaterialPassResource, ResourceArc};
-use rafx::nodes::{
-    FeatureCommandWriter, FeatureSubmitNodes, FramePacket, PrepareJob, RenderFeature,
-    RenderFeatureIndex, RenderJobPrepareContext, RenderView, ViewSubmitNodes,
-};
 
-pub struct TextPrepareJobImpl {
+pub struct PrepareJobImpl {
     text_material_pass: ResourceArc<MaterialPassResource>,
-    extracted_text_data: ExtractedTextData,
+    text_draw_commands: Vec<TextDrawCommand>,
+    font_assets: FnvHashMap<LoadHandle, FontAsset>,
 }
 
-impl TextPrepareJobImpl {
+impl PrepareJobImpl {
     pub(super) fn new(
         text_material_pass: ResourceArc<MaterialPassResource>,
-        extracted_text_data: ExtractedTextData,
+        text_draw_commands: Vec<TextDrawCommand>,
+        font_assets: FnvHashMap<LoadHandle, FontAsset>,
     ) -> Self {
-        TextPrepareJobImpl {
+        PrepareJobImpl {
             text_material_pass,
-            extracted_text_data,
+            text_draw_commands,
+            font_assets,
         }
     }
 }
 
-impl<'a> PrepareJob for TextPrepareJobImpl {
+impl<'a> PrepareJob for PrepareJobImpl {
     fn prepare(
         self: Box<Self>,
         prepare_context: &RenderJobPrepareContext,
         _frame_packet: &FramePacket,
         views: &[RenderView],
     ) -> (Box<dyn FeatureCommandWriter>, FeatureSubmitNodes) {
-        profiling::scope!("Text Prepare");
+        profiling::scope!(prepare_scope);
 
         let dyn_resource_allocator = prepare_context
             .resource_context
@@ -44,10 +46,11 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
         let mut font_atlas_cache = prepare_context
             .render_resources
             .fetch_mut::<FontAtlasCache>();
+
         let draw_vertices_result = font_atlas_cache
             .generate_vertices(
-                &self.extracted_text_data.text_draw_commands,
-                &self.extracted_text_data.font_assets,
+                &self.text_draw_commands,
+                &self.font_assets,
                 &dyn_resource_allocator,
             )
             .unwrap();
@@ -57,6 +60,7 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
             .create_descriptor_set_allocator();
 
         // Get the layouts for both descriptor sets
+
         let per_view_descriptor_set_layout =
             &self.text_material_pass.get_raw().descriptor_set_layouts
                 [shaders::text_vert::PER_VIEW_DATA_DESCRIPTOR_SET_INDEX];
@@ -65,11 +69,16 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
             &self.text_material_pass.get_raw().descriptor_set_layouts
                 [shaders::text_frag::TEX_DESCRIPTOR_SET_INDEX];
 
-        let mut per_font_descriptor_sets = Vec::default();
-        let mut per_view_descriptor_sets = Vec::default();
         let mut submit_nodes = FeatureSubmitNodes::default();
 
-        if !draw_vertices_result.draw_call_metas.is_empty() {
+        let mut writer = Box::new(FeatureCommandWriterImpl::new(
+            self.text_material_pass.clone(),
+            draw_vertices_result.draw_call_metas,
+            draw_vertices_result.image_updates,
+            draw_vertices_result.draw_call_buffer_data.len(),
+        ));
+
+        if !writer.draw_call_metas().is_empty() {
             //
             // Create per-font descriptor sets (i.e. a font atlas texture)
             //
@@ -81,12 +90,11 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
                     )
                     .unwrap();
 
-                per_font_descriptor_sets.push(per_font_descriptor_set.clone());
+                writer.push_per_font_descriptor_set(per_font_descriptor_set);
             }
 
             //
-            // Create per-view descriptor sets (i.e. a projection matrix) and add a submit node per
-            // each view
+            // Create per-view descriptor sets (i.e. a projection matrix)
             //
             for view in views {
                 //
@@ -114,24 +122,14 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
                     )
                     .unwrap();
 
-                // Grow the array if necessary
-                per_view_descriptor_sets.resize(
-                    per_view_descriptor_sets
-                        .len()
-                        .max(view.view_index() as usize + 1),
-                    None,
-                );
-
-                per_view_descriptor_sets[view.view_index() as usize] =
-                    Some(per_view_descriptor_set.clone());
+                writer.push_per_view_descriptor_set(view.view_index(), per_view_descriptor_set);
             }
         }
 
         //
         // Update the vertex buffers
         //
-        let mut draw_call_buffers =
-            Vec::with_capacity(draw_vertices_result.draw_call_buffer_data.len());
+
         for draw_call in draw_vertices_result.draw_call_buffer_data {
             let vertex_buffer = prepare_context
                 .device_context
@@ -159,10 +157,7 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
 
             let index_buffer = dyn_resource_allocator.insert_buffer(index_buffer);
 
-            draw_call_buffers.push(TextDrawCallBuffers {
-                vertex_buffer,
-                index_buffer,
-            })
+            writer.push_buffers(vertex_buffer, index_buffer);
         }
 
         //
@@ -172,33 +167,26 @@ impl<'a> PrepareJob for TextPrepareJobImpl {
         for view in views {
             let mut view_submit_nodes =
                 ViewSubmitNodes::new(self.feature_index(), view.render_phase_mask());
-            for (i, draw_call) in draw_vertices_result.draw_call_metas.iter().enumerate() {
+
+            for (submit_node_id, draw_call) in writer.draw_call_metas().iter().enumerate() {
                 view_submit_nodes.add_submit_node::<UiRenderPhase>(
-                    i as u32,
+                    submit_node_id as u32,
                     0,
                     draw_call.z_position,
                 );
             }
+
             submit_nodes.add_submit_nodes_for_view(view, view_submit_nodes);
         }
-
-        let writer = Box::new(TextCommandWriter {
-            draw_call_buffers,
-            draw_call_metas: draw_vertices_result.draw_call_metas,
-            text_material_pass: self.text_material_pass,
-            per_font_descriptor_sets,
-            per_view_descriptor_sets,
-            image_updates: draw_vertices_result.image_updates,
-        });
 
         (writer, submit_nodes)
     }
 
     fn feature_debug_name(&self) -> &'static str {
-        TextRenderFeature::feature_debug_name()
+        render_feature_debug_name()
     }
 
     fn feature_index(&self) -> RenderFeatureIndex {
-        TextRenderFeature::feature_index()
+        render_feature_index()
     }
 }
