+++ demo/src/features/text/public/mod.rs
@@ -1,0 +1,2 @@
+pub mod text_resource;
+pub use text_resource::*;
